test: add more tests and fixed some bugs in fs_impl
